Hold global configuration via Hanami.configure
