+++ Rakefile
@@ -5,7 +5,7 @@ require 'rspec/core/rake_task'
 namespace :spec do
   RSpec::Core::RakeTask.new(:unit) do |task|
     file_list = FileList['spec/**/*_spec.rb']
-    file_list = file_list.exclude("spec/integration/**/*_spec.rb")
+    file_list = file_list.exclude("spec/{integration,isolation}/**/*_spec.rb")
 
     task.pattern = file_list
   end
