+++ spec/isolation/hanami/configuration_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe "Hanami.configuration" do
+  it "raises error when try to access configuration, but not configured yet" do
+    expect { Hanami.configuration }.to raise_error(RuntimeError, "not configured")
+  end
+end
